js new rent and index done
