+++ app/javascript/packs/application.js
@@ -24,6 +24,7 @@ require("channels")
 
 // External imports
 import "bootstrap";
+import { price } from "../vanilla/price";
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
@@ -31,4 +32,7 @@ import "bootstrap";
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
+  if (document.querySelector(".rent_start_date")) {
+    price();
+  }
 });
