+++ app/views/pages/dashboard.html.erb
@@ -16,7 +16,7 @@
           <%= rent.vinyl.name %><hr>
           There is <%= (rent.end_date - Time.now).fdiv(86400).to_i + 1 %> Days left <hr>
           Owner is <%= rent.vinyl.user.first_name %> <%= rent.vinyl.user.last_name %> <hr>
-          You paid : <%= rent.price %>€
+          You paid : <%= '%.2f' % rent.price %>€
 
         <% end %>
