+++ app/views/pages/home.html.erb
@@ -2,7 +2,7 @@
   <section id="intro">
     <div class="figure-intro">
       <img class="circle-intro" src="https://raw.githubusercontent.com/MartaFagundez/FreeCodeCamp/master/LandingPage/cactus_intro_circle.png" alt="">
-      <img class="img-intro" src="assets/vinyl_green.png" alt="vinyl green">
+      <%= image_tag "vinyl_green.png", alt: "vinyl green", class: "img-intro" %>
     </div>
 
     <h1>Share And Rent <br> all of your <span class="uppercase accent">Vinyls</span> <br>With Our Community</h1>
@@ -19,7 +19,7 @@
       <img class="img-product landscape" src="https://cdn.pixabay.com/photo/2017/08/06/14/29/turntable-2593014_1280.jpg" alt="Vinyl turquoise">
       <img class="img-product portrait" src="https://cdn.pixabay.com/photo/2020/01/31/19/26/plate-4808793_1280.jpg" alt="Vinyl in wooden box">
       <div class="product-quote portrait">
-        <p><span class="uppercase">Vinyl</span> Is The real deal. I've Always felt Like, Until You Buy The <br><span class="accent">Vinyl Record</span>,<br>You Don't Really Own The Album. It's Some Kind of retro romantic Thing From the Past. It is<br><span class="accent"> Still alive.</span></p>
+        <p><span class="uppercase">Vinyl</span> Is The Real Deal. I've Always Felt Like, Until You Buy The <br><span class="accent">Vinyl Record</span>,<br>You Don't Really Own The Album. It's Some Kind of Retro Romantic Thing From The Past. It Is<br><span class="accent"> Still Alive.</span></p>
         <p class="author">Georgito Copito</p>
       </div>
       <img class="img-product landscape" src="https://cdn.pixabay.com/photo/2019/12/28/18/08/records-4725349_1280.jpg" alt="Box full of vinyls">
@@ -35,27 +35,27 @@
 
     <div class="feature-group">
       <div class="feature">
-        <img src="https://raw.githubusercontent.com/MartaFagundez/FreeCodeCamp/master/LandingPage/feature_1.png" alt="">
-        <h3>Premium service</h3>
-        <p>Lorem ipsum color sit amet, consectetur adipiscing elit.</p>
+        <img src="https://cdn.discordapp.com/attachments/722159002712014958/778560806941622312/service2.png" alt="">
+        <h3>Service</h3>
+        <p>Fast and secured.</p>
       </div>
 
       <div class="feature">
-        <img src="https://raw.githubusercontent.com/MartaFagundez/FreeCodeCamp/master/LandingPage/feature_2.png" alt="">
+        <img src="https://cdn.discordapp.com/attachments/722159002712014958/778560788351680512/comu2.png" alt="">
         <h3>Community</h3>
-        <p>Duis aute irure color in reprehenderit in voluptate velit esse.</p>
+        <p>Groovers are all around the world.</p>
       </div>
 
       <div class="feature">
-        <img src="https://raw.githubusercontent.com/MartaFagundez/FreeCodeCamp/master/LandingPage/feature_3.png" alt="">
+        <img src="https://cdn.discordapp.com/attachments/722159002712014958/778560882011013170/vinyl7.png" alt="">
         <h3>Vintage</h3>
-        <p>Excepteur sint occaecat cupidatat non proident, sunt in.</p>
+        <p>Collect them all, they said.</p>
       </div>
     </div>
 
     <div class="explore">
-      <a class="button" href="/vinyls" target="_blank">Explore our Collections</a>
-      <img src="https://raw.githubusercontent.com/MartaFagundez/FreeCodeCamp/master/LandingPage/explore.png" alt="">
+      <a class="button" href="/vinyls">Explore our Collections</a>
+      <img src="https://cdn.discordapp.com/attachments/722159002712014958/778567251741507584/somuchdisk.png" alt="">
     </div>
   </section>
