+++ app/views/shared/_navbar.html.erb
@@ -2,7 +2,7 @@
   <div class="navbar navbar-expand-sm navbar-lewagon">
 
     <a class="navbar-brand" href="/">
-      <img src="https://img2.freepng.fr/20180601/vui/kisspng-logo-brand-font-vinyl-5b10f2e81101d7.5938640415278374160697.jpg" />
+      <%= image_tag "vinyl_green.png" %>
     </a>
 
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
