+++ app/views/vinyls/index.html.erb
@@ -8,6 +8,7 @@
   <ul class="card1-list">
     <h2 class="display-2">Browse</h2>
       <h4>Sort by</h4>
+
       <form class="form-inline">
         <p><%= check_box_tag :quality, class: "cb-quality" %>  Quality</p>
         <p><%= check_box_tag :year, class: "cb-year" %>  Year</p>
