+++ db/seeds.rb
@@ -9,8 +9,8 @@ Rent.destroy_all
 Vinyl.destroy_all
 User.destroy_all
 
-
-user1 = User.create!(email: 'toto@yahoo.fr', password: 'abcdef', password_confirmation: 'abcdef', first_name: "Georgito", last_name: "Yito", address: "84 boulevard massena, Paris")
+user = User.create!(email: 'toto@yahoo.fr', password: 'abcdef', password_confirmation: 'abcdef', first_name: "Georgito", last_name: "Yito", address: "84 boulevard massena, Paris")
+user1 = User.create!(email: 'toto1@yahoo.fr', password: 'abcdef', password_confirmation: 'abcdef', first_name: "Martin", last_name: "Tran", address: "110 avenue de choisy, Paris")
 user2 = User.create!(email: 'toto2@yahoo.fr', password: 'abcdef', password_confirmation: 'abcdef', first_name: "Olivier", last_name: "Calmels", address: "10 rue du mec perdu Paris")
 user3 = User.create!(email: 'toto3@yahoo.fr', password: 'abcdef', password_confirmation: 'abcdef', first_name: "Amal", last_name: "Gouia", address: "10 rue du mec perdu Paris")
 user4 = User.create!(email: 'toto4@yahoo.fr', password: 'abcdef', password_confirmation: 'abcdef', first_name: "Gerom", last_name: "Bretzel", address: "10 rue du mec perdu Paris")
@@ -89,7 +89,7 @@ Vinyl.create(name: "More Miles Per Hour",
 
 Vinyl.create(name: "Nightwatch",
          year:"1978",
-         artist:"Kenny Loggins8",
+         artist:"Kenny Loggins",
         genre: "Rock",
         label: "Columbia",
         quality: "Very Good",
@@ -130,3 +130,24 @@ Vinyl.create(name: "Let It Be",
 #             user: user,
 #             dimension: dimensions1
 #             )
+
+# Vinyl.create(name: "",
+#             year: "",
+#             artist: "",
+#             genre: "",
+#             label: "",
+#             quality: "",
+#             price_per_day: ,
+#             user: user,
+#             dimension: dimensions1
+#             )
+
+
+
+
+
+
+
+
+
+
